Handle part selects of nets that fall of the ends of the identifier.
The Verilog LRM specifies that it is legal to give constant part
selects that are beyond the bounds of the identifier being selected.
But elaboration was flagging that as an error. This patch changes it
to a warning, and handles the cases by generating 'bx bits as needed.

+++ PExpr.h
@@ -387,7 +387,7 @@ class PEIdent : public PExpr {
       NetNet*make_implicit_net_(Design*des, NetScope*scope) const;
 
       bool eval_part_select_(Design*des, NetScope*scope, NetNet*sig,
-			     unsigned&midx, unsigned&lidx) const;
+			     long&midx, long&lidx) const;
       NetNet*process_select_(Design*des, NetScope*scope, NetNet*sig) const;
 
 };
